backend/requirements.txt
# Framework web
Flask==3.0.2
flask-cors==4.0.0

# Base de données
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
Flask-SQLAlchemy==3.1.1

# Scraping
beautifulsoup4==4.12.3
requests==2.31.0
scrapy==2.11.1

# Tests
pytest==8.0.2
pytest-cov==4.1.0

# Utilitaires
python-dotenv==1.0.1
